fix: Use the correct regular expression to parse Cmnd_Alias and other aliases.
Cause: The regex was not taking into consideration that the Cmnd_Alias value
does not have to have spaces on either side of the `=`.  The same is true for
other Alias values.

Consequence: The regex would never terminate, and the role would appear to hang.

Fix: Ensure the regex complies with the eBNF definition of the field from the
sudoers file specification.

Result: The Alias values are parsed correctly.

https://issues.redhat.com/browse/RHEL-106261

Signed-off-by: Rich Megginson <rmeggins@redhat.com>